+++ doc/version_history.rst
@@ -7,10 +7,10 @@ Version History
 ===============
 
 v1.21.0
-_______
+-------
 
 * Add new ``maintel/m1m3/lower_m1m3.py`` sal script and associated files. 
-
+* Add new ``auxtel/offset_ataos.py`` script to offset the ATAOS.
 
 v1.20.1
 -------